fix(inlineerror): adding inline error on empt ystate
